Move enableInternet to PrefManager
